requirement.txt
Django
Djangorestframework
Djangorestframework-simplejwt
pillows
SQlite
Postgresql
